index.ts: remove -linux suffix from userAgent
The userAgent is taken from the version number. However, when the server
does a version check, it considers the suffix to be a different version
and we get the following error:

	The voice server does not support your version of CrewLink.
	Supported versions: 1.2.0,1.2.1

Remove the -linux suffix when setting the userAgent so that the server
recognizes this as a legitimate version of CrewLink.

Currently, the release process works around this by tagging a commit not
on any branch which modifies the package.json version to remove the
-linux suffix. This change obviates that need.